pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentprobe"
version = "0.1.0"
description = "Agent regression testing for production LLM systems"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["llm", "agents", "testing", "evaluation", "regression", "ci"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.25"]
openai = ["openai>=1.0"]
dev = ["pytest>=8.0", "pytest-cov>=4.0", "anthropic>=0.25", "openai>=1.0"]

[project.scripts]
agentprobe = "agentprobe.cli:main"

[project.urls]
Homepage = "https://github.com/FallenOne269/agentprobe"
Issues = "https://github.com/FallenOne269/agentprobe/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentprobe*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
